[Probability] Changed distribution parameter's type to double
Plots are still rendered in float but computations are now in double

+++ apps/probability/distribution/one_parameter_distribution.h
@@ -10,7 +10,7 @@ class OneParameterDistribution : public Distribution {
 public:
   OneParameterDistribution(float parameterValue) : m_parameter1(parameterValue) {}
   int numberOfParameter() override { return 1; }
-  float parameterValueAtIndex(int index) override {
+  double parameterValueAtIndex(int index) override {
     assert(index == 0);
     return m_parameter1;
   }
@@ -19,7 +19,7 @@ public:
     m_parameter1 = f;
   }
 protected:
-  float m_parameter1;
+  double m_parameter1;
 };
 
 }
